Use reagent/with-let to retain the devcard frame (#16)
Because we are using `reagent/with-let` in `show-main` to hold on to the app-db, we need to also use `reagent/with-let` in `show-card` to hold on to the frame. Otherwise if the card re-renders we get a new frame (which the card will use), but the inspector still shows the app-db of the old frame.

On top of that the app-db gets initialized in `show-main`, so if we get a new
frame but show-main still has the old app-db, then the new app-db doesn't get
initialized.