Interoperable Markdown (#12)
Introduce Nextjournal (interoperable) Markdown Parsing 

with api

- `nextjournal.markdown/parse` 
- `nextjournal.markdown/->hiccup` 

The first ingests a markdown string, returns an AST of nodes (clojure maps) with a `:type` and nested `:content`. 

The second function handles configurable transformation of a markdown string into hiccup. We provide default conversions for all nodes but offer a way to overrides single types by providing a function which takes a context and a node and returns hiccup. As a result we don't rely on `(.render MarkdonIt)` any in the markdown viewer, but compose with the hiccup viewer.

This API is usable in clj (via graaljs polyglot library, consuming an esm artifact) as well in cljs, they both share a minimal js source depending on the markdown-it js library. The parsing pass should give us more freedom to introduce new constructs and rely less on markdown-it plugins.

The esm artifact for graal consumption is currently tracked in the repo and accessible from within resources. Babashka tasks for building it were added.

Parsed fenced code blocks carry language and key-value pairs info. 

Add a prototype for zooming into a section (returning a slice of the document).

Distinguish between inline and block images.

Supports a `[[TOC]]` placeholder which is rendered to a nested Table of Contents when producing hiccup. TOC list items have a default click handler which scrolls the referenced heading into view (cljs only).

Supports todo-lists, hashtags and sidenotes.

Adds some basic tests / test runner to the repo.

Bump markdown-it version to 12.2.0

Next steps: 
- add extra delimiters to texmath plugin (as in nextjournal)
- port nextjournal's github-preamble plugin (or handle preamble prior to call `tokenize`)
- make hashtags link href configurable in context

Co-authored-by: Dieter Komendera <dieter@komendera.com>
Co-authored-by: Philippa Markovics <philippa@markovics.com>
Co-authored-by: Martin Kavalar <mk@katercalling.com>